+++ packages/core/nuxt-theme-module/theme/components/CartSidebar.vue
@@ -162,7 +162,6 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@import "~@storefront-ui/vue/styles";
 #cart {
   @include for-desktop {
     & > * {