+++ packages/core/nuxt-theme-module/theme/pages/Category.vue
@@ -446,7 +446,6 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@import "~@storefront-ui/vue/styles";
 #category {
   box-sizing: border-box;
   @include for-desktop {